% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fitting.R
\name{fit_alpha_model}
\alias{fit_alpha_model}
\title{Fit Alpha Using Bayesian Pareto Model with LAI and Breakpoint Corrections}
\usage{
fit_alpha_model(
  filtered_data,
  LAI,
  prior_mean,
  prior_sd,
  stan_model_code = stan_alpha_model,
  iter = 9000,
  warmup = 6000,
  chains = 4,
  cores = 1,
  refresh = 0
)
}
\arguments{
\item{filtered_data}{A list returned by \code{truncate_filter()}, containing the filtered
tree size data frame (\code{bayesian_data}), the final breakpoint (\code{final_breakpoint}),
and the log-scaled KDE data frame (\code{kerneldens_logtransform}).}

\item{LAI}{A numeric value for site-level Leaf Area Index. The function assumes your LAI value is on
a 1-10 scale and will divide by 10 to get LAI between 0 and 1 (e.g., if you have an LAI of 5 on a
scale of 1-10, enter 5, while if you have a 50 on a scale of 1-100, enter 5).}

\item{prior_mean}{Prior mean for the alpha parameter.}

\item{prior_sd}{Prior standard deviation for the alpha parameter.}

\item{stan_model_code}{Character string of a Stan model (default = \code{stan_alpha_model}).}

\item{iter}{Number of Stan iterations (default = 9000).}

\item{warmup}{Number of warmup iterations (default = 6000).}

\item{chains}{Number of chains (default = 4).}

\item{cores}{Number of CPU cores to use for parallel chains (default = 1).}

\item{refresh}{Frequency of Stan progress output (default = 0, i.e., silent).}
}
\value{
A list with:
\describe{
\item{posterior_summary}{Data frame of posterior summaries for alpha with R2.}
\item{stan_fit}{Stan fit object.}
\item{breakpoint_norm}{Normalized breakpoint value passed to downstream functions.}
\item{LAI_norm}{Normalized LAI value passed to downstream functions.}
\item{bayesian_data}{The filtered data frame used for model fitting.}
}
}
\description{
This function fits a Pareto model to tree size data using a Bayesian approach with Stan,
accounting for reduced visibility of small trees and site-level LAI corrections. Note that running
this function will require rstan, which itself needs Rtools. Rtools is not a CRAN package and as far
as we know must be installed directly. Try this link (https://cran.r-project.org/bin/windows/Rtools/)
}
\examples{
\dontrun{
data("harv_data_sample")
df <- harv_data_sample[harv_data_sample$IDhectbest == 1, "dbh", drop = FALSE]
kde_output   <- potential_break(df)
trunc_output <- truncate_filter(kde_output)
fit <- fit_alpha_model(trunc_output, LAI = 5.4, prior_mean = 1.4, prior_sd = 0.3)
}
}
